refactor: split upload endpoints
