+++ functions/src/shared/env.ts
@@ -1,0 +1,11 @@
+import { HttpError } from './httpError'
+
+export function mustGetEnv(name: string): string {
+  const value = process.env[name]
+
+  if (!value) {
+    throw new HttpError(500, `Missing required environment variable: ${name}`)
+  }
+
+  return value
+}
