+++ functions/src/shared/httpError.ts
@@ -1,0 +1,8 @@
+export class HttpError extends Error {
+  constructor(
+    public readonly statusCode: number,
+    message: string
+  ) {
+    super(message)
+  }
+}
