+++ functions/src/shared/validation.ts
@@ -1,0 +1,9 @@
+import { HttpError } from './httpError'
+
+export function validateRequiredString(value: unknown, fieldName: string): string {
+  if (typeof value !== 'string' || !value.trim()) {
+    throw new HttpError(400, `${fieldName} is required.`)
+  }
+
+  return value.trim()
+}
